rewrite to cats effect

+++ src/main/scala/net/michalp/identicon4s/Hashing.scala
@@ -27,7 +27,7 @@ object Hashing {
 
   def instance[F[_]: Applicative] = new Hashing[F] {
     def hash(input: String): F[Long] =
-      Applicative[F].pure(input.hashCode())
+      Applicative[F].pure(input.hashCode().toLong)
   }
 
 }
